+++ front/src/helpers/scrollUtils.js
@@ -1,0 +1,11 @@
+export const scrollToSection = (event, sectionId) => {
+  event.preventDefault();
+  const section = document.getElementById(sectionId);
+  if (section) {
+    section.scrollIntoView({
+      behavior: 'smooth',
+      block: 'start',
+    });
+  }
+};
+
